packages-ignored/hop3-web/pyproject.toml
[project]
name = "hop3-web"
version = "0.2.3"

[tool.poetry]
name = "hop3-web"
version = "0.2.2"
description = "Simple PaaS - Deploy and manage web applications on a single server"
authors = ["Stefane Fermigier <sf@abilian.com>"]
readme = "README.md"

packages = [{ include = "hop3_web", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"

attrs = "*"
click = "^8.1.7"
cleez = "*"
pluggy = "^1.4.0"
svcs = "^24.1.0"

sqlalchemy = "^2.0"
sqlalchemy-serializer = ">= 1.4"
advanced-alchemy = ">= 0.21"
psycopg2 = "^2.9"

# System introspection
psutil = ">= 6.0"
py-cpuinfo = ">= 9.0"

# Remove later
devtools = ">= 0.12"
snoop = ">= 0.6"

# "main" / web app
flask = "^3.0"
flask-super = "^0.2"
sqlalchemy-utils = "^0.41"
flask-security-too = "^5.4"
flask-sqlalchemy = "^3.1"
flask-babel = "^4.0"
flask-htmx = "^0.4"
flask-mail = "^0.10"
flask-migrate = "^4.0"
flask-talisman = "^1.1"
flask-vite = "^0.5"
sentry-sdk = "^2.12"

[tool.poetry.group.dev.dependencies]
abilian-devtools = "^0.7"
httpx = "^0.27"
pytest = "^8.0"
python-dotenv = "^1.0"
reuse = "^4"
infer-types = "^1.0"

pyinfra = "^3"

types-paramiko = "^3.4"
types-psycopg2 = "^2.9"
flask-debugtoolbar = "^0.15"
msgspec = "^0.18"
watchfiles = ">= 0.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pyright]
exclude = ["src/hop3/util/templating.py"]
